Updated version to 0.10.2, bug fixes & improvements
  * Fixed a bug in setStyle() when using BL_CONTEXT_STYLE_TRANSFORM_MODE_NONE
    (identity transform type was not set properly, so the API doesn't respect
    that)
  * Fixed a bug in BLStrokeOptions in which dashArray was deallocated twice
    (added new tests that have uncovered this issue)
  * Fixed some Windows compilation warnings about placement new/free
  * Added a new API to get a final transformation matrix from BLContext - this
    allows to use that transformation matrix to apply a custom transform to a
    pattern and then to use BL_CONTEXT_STYLE_TRANSFORM_MODE_NONE to bypass any
    other transformations that would be done by the rendering context
  * Exposed low-level stroking interface that strokes to a sink - experimental,
    only exposed as C-API for now, before it stabilizes
  * Added more BLContext tests that verify fill and stroke options
  * Added gradient dithering to portable pipelines
  * Updated version to 0.10.2

+++ test/bl_generator.h
@@ -9,7 +9,6 @@
 #include <mutex>
 #include <random>
 #include <string>
-#include <string_view>
 #include <thread>
 #include <vector>
 
@@ -560,7 +559,7 @@ public:
     occupied1.resize(bucketCount1);
 
     {
-      for (size_t i = 0; i < _size; i++) {
+      for (uint32_t i = 0; i < _size; i++) {
         uint32_t index = (_values[i] * _hf.params[0].multiplier) >> _hf.params[0].shift;
         if (occupied1.hasBit(index)) {
           remainingPairs.push_back(std::pair<uint32_t, uint32_t>(_values[i], i));
